my_django/requirements.txt
amqp==2.5.2
asgiref==3.2.7
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.5
django-concurrency==2.1.1
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-nested-routers==0.91
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pydotplus==2.0.2
pygraphviz==1.5
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
sh==1.12.14
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.9
whitenoise==5.0.1
zipp==3.1.0
